Add authentication middleware
A user of the Heimdall API can authenticate with either an API key or a
session token stored within the heimdall_sessionToken cookie. If both
are passed, the API key will be checked first, then the cookie will be
checked. At least one valid method of authentication must be passed for
the request to succeed.