Add setup mode flag
When Heimdall is first run, there will not be any users or clients.
Since an authenticated user or client is required to create other users
and clients, the initial resource cannot be made. This change provides a
workaround by introducing the `setup-mode` flag. Currently, this flag
just disables auth checks for all endpoints, but in the future it could
be used for other intial setup behavior changes.